Added PodiumUsers.kt RankingsUserList.kt, created function to create fake user data for rankings screen

+++ app/src/main/java/com/example/bearvbull/data/users/UserAccountInformation.kt
@@ -3,7 +3,7 @@ package com.example.bearvbull.data.users
 data class UserAccountInformation(
     val userId : String,
     val userName: String,
-    val userBalance: Double,
+    val userBalance: Long,
     val profileImage: Int,
     val rank: Int
 )
